glslang: Fix over 100 warnings from MSVC warning level 4.
git-svn-id: https://cvs.khronos.org/svn/repos/ogl/trunk/ecosystem/public/sdk/tools/glslang@31202 e7fa87d3-cd2b-0410-9028-fcbf551c1848

+++ StandAlone/StandAlone.cpp
@@ -938,7 +938,7 @@ char** ReadFileData(const char* fileName)
         printf("Error allocating memory\n");
         return 0;
     }
-    if (fread(fdata,1,count, in)!=count) {
+    if ((int)fread(fdata,1,count, in) != count) {
             printf("Error reading input file: %s\n", fileName);
             return 0;
     }
